+++ starters/gatsby-starter-catalyst-sanity/sanity-studio/schemas/post.js
@@ -31,28 +31,6 @@ export default {
       validation: (Rule) => Rule.required(),
     },
     {
-      title: "Author",
-      name: "author",
-      type: "reference",
-      to: { type: "author" },
-      validation: (Rule) => Rule.required(),
-    },
-    {
-      name: "featuredImage",
-      title: "Featured Image",
-      description: "A featured image for your post.",
-      type: "figure",
-      validation: (Rule) => Rule.required(),
-    },
-    {
-      name: "excerpt",
-      title: "Post Excerpt",
-      description:
-        "Used to show a summary of your post in the posts index page.",
-      type: "excerptBlockContent",
-      validation: (Rule) => Rule.required(),
-    },
-    {
       name: "body",
       title: "Post Content",
       type: "blockContent",